rename Core/Analytics to Core/DolphinAnalytics
